Move grid functionality to separate component
