+++ src/app/puzzle-editing/puzzle-editing.component.html
@@ -2,35 +2,43 @@
   <div class="row">
     <div class="hstack gap-3">
       <div class="btn-group" role="group" aria-label="Edit Mode">
-        <button type="button" class="btn btn-primary" [ngClass]="{ 'active': editMode == 0 }" (click)="editMode = 0">Answers</button>
-        <button type="button" class="btn btn-primary" [ngClass]="{ 'active': editMode == 1 }" (click)="editMode = 1">Spacers</button>
-        <button type="button" class="btn btn-primary" [ngClass]="{ 'active': editMode == 2 }" (click)="editMode = 2">Circles</button>
-        <button type="button" class="btn btn-primary" [ngClass]="{ 'active': editMode == 3 }" (click)="editMode = 3">Shades</button>
+        <button type="button" class="btn btn-primary" [ngClass]="{ 'active': editMode == 0 }" (click)="editMode = 0; onUpdateConfig()">
+          Answers
+        </button>
+        <button type="button" class="btn btn-primary" [ngClass]="{ 'active': editMode == 1 }" (click)="editMode = 1; onUpdateConfig()">
+          Spacers
+        </button>
+        <button type="button" class="btn btn-primary" [ngClass]="{ 'active': editMode == 2 }" (click)="editMode = 2; onUpdateConfig()">
+          Circles
+        </button>
+        <button type="button" class="btn btn-primary" [ngClass]="{ 'active': editMode == 3 }" (click)="editMode = 3; onUpdateConfig()">
+          Shades
+        </button>
       </div>
       <div class="btn-group" role="group" aria-label="Highlight Mode">
         <button
           type="button"
           class="btn btn-info"
           [ngClass]="{ 'active': highlightMode == 0, 'disabled': editMode != 0 }"
-          (click)="highlightMode = 0">
+          (click)="highlightMode = 0; onUpdateConfig()">
           Across
         </button>
         <button
           type="button"
           class="btn btn-info"
           [ngClass]="{ 'active': highlightMode == 1, 'disabled': editMode != 0 }"
-          (click)="highlightMode = 1">
+          (click)="highlightMode = 1; onUpdateConfig()">
           Down
         </button>
         <button
           type="button"
           class="btn btn-info"
           [ngClass]="{ 'active': highlightMode == 2, 'disabled': editMode != 0 }"
-          (click)="highlightMode = 2">
+          (click)="highlightMode = 2; onUpdateConfig()">
           Intersect
         </button>
       </div>
-      <button class="btn btn-secondary" type="submit" (click)="answersHidden = !answersHidden">
+      <button class="btn btn-secondary" type="submit" (click)="answersHidden = !answersHidden; onUpdateConfig()">
         <span *ngIf="answersHidden">Show</span><span *ngIf="!answersHidden">Hide</span> Answers
       </button>
       <button class="ms-auto btn btn-success" type="submit" (click)="onSave()">Save</button>
@@ -39,36 +47,7 @@
   </div>
   <div class="row mt-4" *ngIf="puzzleLoaded">
     <div class="col">
-      <table
-        #table
-        class="table table-bordered"
-        [ngStyle]="{ 'width.px': numCols * squareHeight }"
-        (keydown)="onKeyDown($event)"
-        tabindex="0">
-        <tbody>
-          <tr *ngFor="let row of [].constructor(numRows); index as i">
-            <td
-              class="square"
-              *ngFor="let square of getPuzzleRow(i * numCols); index as j"
-              [ngClass]="{
-                'spacer': square.type == 1,
-                'shaded': square.overlay == 2,
-                'selected': square.index == selectedIndex,
-                'highlighted': isHighlighted(square)
-              }"
-              (click)="onClickSquare(square.index)">
-              <div class="square-contents" [ngStyle]="{ 'height.px': squareHeight, 'width.px': squareHeight }">
-                <div class="number" *ngIf="square.boxNum != -1">{{ square.boxNum }}</div>
-                <div class="circle" [ngStyle]="{ 'border': square.overlay == 1 ? 'solid' : 'none' }">
-                  <div class="value text-center" [ngStyle]="{ 'display': answersHidden ? 'none' : 'block' }">
-                    <span>{{ square.value }}</span>
-                  </div>
-                </div>
-              </div>
-            </td>
-          </tr>
-        </tbody>
-      </table>
+      <app-grid [config$]="puzzleConfig$"></app-grid>
     </div>
     <div class="col">
       <app-selected-clue></app-selected-clue>
